requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pycparser==2.20
pylint==2.9.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
